// client/src/app/expected-result-comparison/models/ExpectedVsTimeSeriesRequestConfigModel.ts
import { ModelDictionaryComponent } from '@client:common/topology';
export interface ExpectedVsTimeSeriesRequestConfigModel {
  lineName: string | '';
  userSelectedSimulationId: number | null;
  componentType: string;
  useMagnitude: boolean | null;
  useAngle: boolean | null;
  component: ModelDictionaryComponent[];
  firstSimulationId: number | null;
}